Adapting it to SODA submission
In the SODA guidelines the first page must contain only the abstract, and the reviewers are only required to read the next 10 pages after the abstract, but the submission can have any number of pages. Hence I tried to clean the first 11 pages as much as possible, sending a bunch of stuff to appendices and leaving only the most interesting bits.